Refactoring course contollers out of home controller
