+++ src/controllers/homeController.ts
@@ -1,10 +1,9 @@
-import {NextFunction, Request, Response} from 'express'
+import {Request, Response} from 'express'
 import {LearningCatalogue} from '../learning-catalogue'
 import {Course} from '../learning-catalogue/model/course'
 import {DefaultPageResults} from '../learning-catalogue/model/defaultPageResults'
 
 import * as log4js from 'log4js'
-import {CourseRequest} from '../extended'
 
 export class HomeController {
 	logger = log4js.getLogger('controllers/homeController')
@@ -36,58 +35,6 @@ export class HomeController {
 			response.render('page/index', {
 				pageResults,
 			})
-		}
-	}
-
-	public courseOverview() {
-		return async (request: Request, response: Response) => {
-			const req = request as CourseRequest
-
-			const course = req.course
-
-			response.render(`page/course`, {course})
-		}
-	}
-
-	public addModule() {
-		return async (request: Request, response: Response) => {
-			response.render(`page/add-module`)
-		}
-	}
-
-	public addModuleBlog() {
-		return async (request: Request, response: Response) => {
-			response.render(`page/add-module-blog`)
-		}
-	}
-
-	public coursePreview() {
-		return async (request: Request, response: Response) => {
-			const req = request as CourseRequest
-
-			const course = req.course
-
-			response.render('page/course-preview', {course})
-		}
-	}
-
-	public loadCourse() {
-		const self = this
-
-		return async (
-			request: Request,
-			response: Response,
-			next: NextFunction
-		) => {
-			const req = request as CourseRequest
-			const courseId: string = req.params.courseId
-			const course = await self.learningCatalogue.getCourse(courseId)
-			if (course) {
-				req.course = course
-				next()
-			} else {
-				response.sendStatus(404)
-			}
 		}
 	}
 }
