+++ test/unit/controllers/courseControllerTest.ts
@@ -45,10 +45,9 @@ describe('Course Controller Tests', function() {
 		const request: Request = mockReq()
 		const response: Response = mockRes()
 
-		const req = request as CourseRequest
-		req.course = course
+		learningCatalogue.getCourse = sinon.stub().returns(course)
 
-		await courseOverview(req, response)
+		await courseOverview(request, response)
 
 		expect(response.render).to.have.been.calledOnceWith('page/course', {
 			course,
